review, chnage pass, wishlist, cancel order is done

+++ app/src/main/java/com/teamx/hatlyUser/ui/fragments/foods/foodsShopPreview/modelShopHome/Option.kt
@@ -1,10 +1,9 @@
 package com.teamx.hatlyUser.ui.fragments.foods.foodsShopPreview.modelShopHome
 import androidx.annotation.Keep
-
 @Keep
 data class Option(
     val _id: String,
     val name: String,
-    val price: Double,
-    val salePrice: Double
+    val price: Int,
+    val salePrice: Int
 )
